+++ src/canister/src/logic/websocket_logic.rs
@@ -8,7 +8,7 @@ use ic_websocket_cdk::{
     WsInitParams,
 };
 
-use crate::storage::{StorageQueryable, UserNotificationStore};
+use crate::storage::{RewardBufferStore, StorageQueryable, UserNotificationStore};
 
 thread_local! {
    pub static CONNECTED_CLIENTS: RefCell<HashMap<Principal, u64>> = RefCell::new(HashMap::new());
@@ -32,6 +32,9 @@ impl Websocket {
 
     pub fn on_open(args: OnOpenCallbackArgs) {
         Self::add_connected_to_clients(args.client_principal);
+
+        RewardBufferStore::notify_active_user(args.client_principal);
+
         let (_, notification_ids) = UserNotificationStore::get(args.client_principal)
             .unwrap_or((args.client_principal, UserNotifications::new()));
         Self::send_message(