+++ src/canister/src/storage/mod.rs
@@ -15,11 +15,14 @@ mod member_storage;
 mod notification_storage;
 mod profile_storage;
 mod report_storage;
+pub mod reward_canister_storage;
+pub mod reward_storage;
 pub mod storage_api;
 mod topic_storage;
 mod user_notification_storage;
 
 // Re-export stores
+
 pub use attendee_storage::AttendeeStore;
 pub use boosted_storage::BoostedStore;
 pub use cell_api::{CellStorage, CellStorageRef};
@@ -43,6 +46,7 @@ pub use user_notification_storage::UserNotificationStore;
 
 pub use id_storage::*;
 pub use logger_storage::LoggerStore;
+pub use reward_storage::{RewardBufferStore, RewardTimerStore};
 
 // Re-export constants
 pub use logger_storage::LOGIN_EVENT;